// jara-project-4.0/react-test/app/utils/interface.tsx
//国,都道府県マスターの共通インターフェース
export interface masterDataResponse {
    id: number;
    dispNum: number;
    name: string;
    code: string;
}

//国,都道府県マスターの共通インターフェース
export interface listDataResponse {
    id: number;
    name: string;
    code: string;
}